email_negotiation/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "openenv-email-negotiation"
version = "0.1.0"
description = "Enterprise email negotiation environment for RL training"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.3",
    "openai>=1.0.0",
    "numpy>=1.24.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
]

[project.scripts]
server = "server.app:main"

[tool.hatch.build.targets.wheel]
packages = ["."]